Organization Service Rest

+++ uengine-bill-web/src/main/java/org/uengine/garuda/web/organization/OrganizationServiceImpl.java
@@ -1,16 +1,19 @@
 package org.uengine.garuda.web.organization;
 
 import org.opencloudengine.garuda.client.model.OauthUser;
-import org.opencloudengine.garuda.util.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.uengine.garuda.common.exception.ServiceException;
+import org.uengine.garuda.killbill.KBRepository;
 import org.uengine.garuda.killbill.KBServiceFactory;
 import org.uengine.garuda.killbill.api.model.Tenant;
 import org.uengine.garuda.model.Authority;
 import org.uengine.garuda.model.Organization;
 import org.uengine.garuda.model.OrganizationEmail;
+import org.uengine.garuda.util.StringUtils;
 import org.uengine.garuda.web.system.UserService;
 
 import javax.servlet.http.HttpServletRequest;
@@ -27,36 +30,67 @@ public class OrganizationServiceImpl implements OrganizationService {
     private OrganizationRepository organizationRepository;
 
     @Autowired
+    private KBRepository kbRepository;
+
+    @Autowired
     private UserService userService;
 
     @Autowired
     private KBServiceFactory killbillServiceFactory;
 
+    private Logger logger = LoggerFactory.getLogger(OrganizationService.class);
+
     /**
      * 조직을 생성한다.
+     *
      * @param organization Organization
      * @return
      */
     @Override
-    public Organization createOrganization(Organization organization) {
+    public Organization createOrganization(Organization organization, OauthUser oauthUser) {
+
+        logger.info("Start to create organization {}", organization.getName());
+
         String name = organization.getName();
-        if(StringUtils.isEmpty(name)){
+        if (StringUtils.isEmpty(name)) {
             throw new ServiceException("organization name is required");
         }
+
+        //킬빌 테넌트를 생성한다.
         Tenant tenant = new Tenant();
         String key = UUID.randomUUID().toString();
         tenant.setApiSecret(key);
         tenant.setApiKey(key);
         tenant.setExternalKey(key);
+        Tenant createdTenant = killbillServiceFactory.apiClient().tenantApi().createTenant(tenant);
 
-        Tenant created = killbillServiceFactory.apiClient().tenantApi().createTenant(tenant);
+        //조직을 생성한다.
+        organization.setTenant_id(createdTenant.getTenantId());
+        organization.setTenant_api_key(tenant.getApiKey());
+        organization.setTenant_api_secret(tenant.getApiSecret());
+        organization.setTenant_external_key(tenant.getExternalKey());
+        Organization createdOrganization = organizationRepository.insert(organization);
 
-        organization.setTenant_id(created.getTenantId());
-        organization.setTenant_api_key(created.getApiKey());
-        organization.setTenant_api_secret(created.getApiSecret());
-        organization.setTenant_external_key(created.getExternalKey());
+        //요청자를 ADMIN 으로 Authority 를 생성한다.
+        Authority authority = new Authority();
+        authority.setOrganization_id(createdOrganization.getId());
+        authority.setUser_id(oauthUser.get_id());
+        authority.setUser_name(oauthUser.getUserName());
+        authority.setRole(OrganizationRole.ADMIN);
+        organizationRepository.insertAuthority(authority);
 
-        return organizationRepository.insert(organization);
+        //요청자의 이메일로 OrganizationEmail 을 생성한다.
+        if (oauthUser.containsKey("email")) {
+            OrganizationEmail organizationEmail = new OrganizationEmail();
+            organizationEmail.setOrganization_id(createdOrganization.getId());
+            organizationEmail.setEmail(oauthUser.get("email").toString());
+            organizationEmail.setIs_active("Y");
+            organizationRepository.insertOrganizationEmail(organizationEmail);
+        }
+
+        logger.info("Success to create organization {}", organization.getName());
+
+        return createdOrganization;
     }
 
     @Override
@@ -66,6 +100,14 @@ public class OrganizationServiceImpl implements OrganizationService {
 
     @Override
     public int deleteOrganization(String id) {
+        Organization organization = this.selectById(id);
+
+        //구독건이 남아있다면 삭제 불가능
+        Long count = kbRepository.subscriptionCountByTenantId(organization.getTenant_id());
+        if (count != null && count > 0) {
+            logger.error("Failed to delete organization {}", id);
+            throw new ServiceException("One or more subscriptions are exist. : " + count);
+        }
         return organizationRepository.delete(id);
     }
 
@@ -152,6 +194,7 @@ public class OrganizationServiceImpl implements OrganizationService {
 
     /**
      * http 헤더의 토큰값과 organization id 로 접근 가능 롤을 판별한다.
+     *
      * @param request
      * @return
      */
@@ -165,41 +208,41 @@ public class OrganizationServiceImpl implements OrganizationService {
 
         orgRole.setRole(role);
 
-        try{
+        try {
             //토큰값이 없다면 fail
-            if(StringUtils.isEmpty(token)){
+            if (StringUtils.isEmpty(token)) {
                 orgRole.setAccept(false);
                 return orgRole;
-            }else{
+            } else {
                 orgRole.setToken(token);
             }
 
             //토큰 정보가 없다면 fail
             Map tokenInfo = userService.tokenInfo(token);
-            if(tokenInfo == null){
+            if (tokenInfo == null) {
                 orgRole.setAccept(false);
                 return orgRole;
-            }else{
+            } else {
                 orgRole.setToken(token);
             }
 
             //유저 정보를 얻어오지 못하면 fail
-            String user_id = ((Map)tokenInfo.get("context")).get("userId").toString();
+            String user_id = ((Map) tokenInfo.get("context")).get("userId").toString();
             OauthUser user = userService.selectByUserId(user_id);
-            if(user == null){
+            if (user == null) {
                 orgRole.setAccept(false);
                 return orgRole;
-            }else{
+            } else {
                 orgRole.setOauthUser(user);
             }
 
             //organization_id 가 없고 ANY 일 경우 success 리턴
             //organization_id 가 없고 그 외의 경우 fail
-            if(StringUtils.isEmpty(organization_id)){
-                if(OrganizationRole.ANY.equals(role)){
+            if (StringUtils.isEmpty(organization_id)) {
+                if (OrganizationRole.ANY.equals(role)) {
                     orgRole.setAccept(true);
                     return orgRole;
-                }else{
+                } else {
                     orgRole.setAccept(false);
                     return orgRole;
                 }
@@ -207,44 +250,64 @@ public class OrganizationServiceImpl implements OrganizationService {
 
             //조직이 없을 경우 리턴
             Organization organization = organizationRepository.selectById(organization_id);
-            if(organization == null){
+            if (organization == null) {
                 orgRole.setAccept(false);
                 return orgRole;
-            }else{
+            } else {
                 orgRole.setOrganization(organization);
             }
 
             Authority authority = organizationRepository.selectAuthorityByUserIdAndOrganizationId(user_id, organization_id);
             //authority 가 없을 경우 fail
-            if (authority == null){
+            if (authority == null) {
                 orgRole.setAccept(false);
                 return orgRole;
             }
 
             //authority 와 role 규칙이 틀리면 fail
-            if(OrganizationRole.ADMIN.equals(role)){
-                if(OrganizationRole.ADMIN.equals(authority.getRole())){
+            if (OrganizationRole.ADMIN.equals(role)) {
+                if (OrganizationRole.ADMIN.equals(authority.getRole())) {
                     orgRole.setAccept(true);
                     return orgRole;
                 }
             }
-            if(OrganizationRole.MEMBER.equals(role)){
-                if(OrganizationRole.ADMIN.equals(authority.getRole()) || OrganizationRole.MEMBER.equals(authority.getRole())){
+            if (OrganizationRole.MEMBER.equals(role)) {
+                if (OrganizationRole.ADMIN.equals(authority.getRole()) || OrganizationRole.MEMBER.equals(authority.getRole())) {
                     orgRole.setAccept(true);
                     return orgRole;
                 }
             }
-            if(OrganizationRole.ANY.equals(role)){
+            if (OrganizationRole.ANY.equals(role)) {
                 orgRole.setAccept(true);
                 return orgRole;
             }
 
             orgRole.setAccept(false);
             return orgRole;
-        }catch (Exception ex){
+        } catch (Exception ex) {
             ex.printStackTrace();
             orgRole.setAccept(false);
             return orgRole;
         }
+    }
+
+    @Override
+    public OauthUser getUserFromRequest(HttpServletRequest request) {
+        Map<String, String> headers = this.getHeaders(request);
+        String token = headers.get("authorization");
+
+        if (StringUtils.isEmpty(token)) {
+            return null;
+        }
+
+        Map tokenInfo = userService.tokenInfo(token);
+        if (tokenInfo == null) {
+            return null;
+        }
+
+        //유저 정보를 얻어오지 못하면 fail
+        String user_id = ((Map) tokenInfo.get("context")).get("userId").toString();
+        OauthUser user = userService.selectByUserId(user_id);
+        return user;
     }
 }
